app/requirements.txt
click==7.1.2
fastapi==0.60.1
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
pydantic==1.6.1
python-dotenv==0.14.0
starlette==0.13.6
uvicorn==0.11.7
uvloop==0.14.0
websockets==8.1
python-dateutil
six
bitarray
thriftpy2
impyla
aiofiles
jinja2
python-multipart
aiofiles